requirements.txt
pandas==1.2.1
imutils==0.5.3
numpy==1.18.2
opencv_python==4.2.0.34
Pillow==8.1.0
scikit_learn==0.24.1
